Change datatype from float to double

+++ include/linear_interpolator.h
@@ -3,15 +3,15 @@
 
 
 struct LinearInterpolator {
-    LinearInterpolator(float input_min, float input_max, float output_min, float output_max);
+    LinearInterpolator(double input_min, double input_max, double output_min, double output_max);
     virtual ~LinearInterpolator();
 
-    float map(float value) const;
+    double map(double value) const;
 
-    float x1;
-    float x2;
-    float y1;
-    float y2;
+    double x1;
+    double x2;
+    double y1;
+    double y2;
 
 
 };
